pyproject.toml
[tool.poetry]
name = "bot"
version = "0.1.0"
description = ""
authors = ["Nastya <annika-purtova@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = {extras = ["socks"], version = "^13.15"}
pydantic = "^1.10.2"
requests = "^2.28.1"


[tool.poetry.group.dev.dependencies]
mypy = "^0.991"
wemake-python-styleguide = "^0.17.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
